Remove ImageSegmenterHelper.kt
